+++ planner/cardinality/trace.go
@@ -21,10 +21,15 @@ import (
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/parser/ast"
 	"github.com/pingcap/tidb/parser/format"
+	"github.com/pingcap/tidb/parser/model"
+	"github.com/pingcap/tidb/planner/util/debugtrace"
 	"github.com/pingcap/tidb/sessionctx"
+	"github.com/pingcap/tidb/sessionctx/stmtctx"
+	"github.com/pingcap/tidb/statistics"
 	driver "github.com/pingcap/tidb/types/parser_driver"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/logutil"
+	"github.com/pingcap/tidb/util/ranger"
 	"github.com/pingcap/tidb/util/tracing"
 	"go.uber.org/zap"
 )
@@ -107,4 +112,125 @@ func exprToString(e expression.Expression) (string, error) {
 		return buffer.String(), nil
 	}
 	return "", errors.New("unexpected type of Expression")
+}
+
+/*
+ Below is debug trace for GetRowCountByXXX().
+*/
+
+type getRowCountInput struct {
+	Ranges []string
+	ID     int64
+}
+
+func debugTraceGetRowCountInput(
+	s sessionctx.Context,
+	id int64,
+	ranges ranger.Ranges,
+) {
+	root := debugtrace.GetOrInitDebugTraceRoot(s)
+	newCtx := &getRowCountInput{
+		ID:     id,
+		Ranges: make([]string, len(ranges)),
+	}
+	for i, r := range ranges {
+		newCtx.Ranges[i] = r.String()
+	}
+	root.AppendStepToCurrentContext(newCtx)
+}
+
+// GetTblInfoForUsedStatsByPhysicalID get table name, partition name and TableInfo that will be used to record used stats.
+var GetTblInfoForUsedStatsByPhysicalID func(sctx sessionctx.Context, id int64) (fullName string, tblInfo *model.TableInfo)
+
+// recordUsedItemStatsStatus only records un-FullLoad item load status during user query
+func recordUsedItemStatsStatus(sctx sessionctx.Context, stats interface{}, tableID, id int64) {
+	// Sometimes we try to use stats on _tidb_rowid (id == -1), which must be empty, we ignore this case here.
+	if id <= 0 {
+		return
+	}
+	var isIndex, missing bool
+	var loadStatus *statistics.StatsLoadedStatus
+	switch x := stats.(type) {
+	case *statistics.Column:
+		isIndex = false
+		if x == nil {
+			missing = true
+		} else {
+			loadStatus = &x.StatsLoadedStatus
+		}
+	case *statistics.Index:
+		isIndex = true
+		if x == nil {
+			missing = true
+		} else {
+			loadStatus = &x.StatsLoadedStatus
+		}
+	}
+
+	// no need to record
+	if !missing && loadStatus.IsFullLoad() {
+		return
+	}
+
+	// need to record
+	statsRecord := sctx.GetSessionVars().StmtCtx.GetUsedStatsInfo(true)
+	if statsRecord[tableID] == nil {
+		name, tblInfo := GetTblInfoForUsedStatsByPhysicalID(sctx, tableID)
+		statsRecord[tableID] = &stmtctx.UsedStatsInfoForTable{
+			Name:    name,
+			TblInfo: tblInfo,
+		}
+	}
+	recordForTbl := statsRecord[tableID]
+
+	var recordForColOrIdx map[int64]string
+	if isIndex {
+		if recordForTbl.IndexStatsLoadStatus == nil {
+			recordForTbl.IndexStatsLoadStatus = make(map[int64]string, 1)
+		}
+		recordForColOrIdx = recordForTbl.IndexStatsLoadStatus
+	} else {
+		if recordForTbl.ColumnStatsLoadStatus == nil {
+			recordForTbl.ColumnStatsLoadStatus = make(map[int64]string, 1)
+		}
+		recordForColOrIdx = recordForTbl.ColumnStatsLoadStatus
+	}
+
+	if missing {
+		recordForColOrIdx[id] = "missing"
+		return
+	}
+	recordForColOrIdx[id] = loadStatus.StatusToString()
+}
+
+// CETraceRange appends a list of ranges and related information into CE trace
+func CETraceRange(sctx sessionctx.Context, tableID int64, colNames []string, ranges []*ranger.Range, tp string, rowCount uint64) {
+	sc := sctx.GetSessionVars().StmtCtx
+	allPoint := true
+	for _, ran := range ranges {
+		if !ran.IsPointNullable(sctx) {
+			allPoint = false
+			break
+		}
+	}
+	if allPoint {
+		tp = tp + "-Point"
+	} else {
+		tp = tp + "-Range"
+	}
+	expr, err := ranger.RangesToString(sc, ranges, colNames)
+	if err != nil {
+		logutil.BgLogger().Debug("Failed to trace CE of ranges", zap.String("category", "OptimizerTrace"), zap.Error(err))
+	}
+	// We don't need to record meaningless expressions.
+	if expr == "" || expr == "true" || expr == "false" {
+		return
+	}
+	ceRecord := tracing.CETraceRecord{
+		TableID:  tableID,
+		Type:     tp,
+		Expr:     expr,
+		RowCount: rowCount,
+	}
+	sc.OptimizerCETrace = append(sc.OptimizerCETrace, &ceRecord)
 }
