+++ statistics/BUILD.bazel
@@ -49,7 +49,6 @@ go_library(
         "//util/memory",
         "//util/ranger",
         "//util/sqlexec",
-        "//util/tracing",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_tipb//go-tipb",
